fix: build to use deno scripts
