--- scripts/build_all
@@ -1,11 +1,0 @@
-#!/bin/bash
-
-# This script only exists to run Deno in different OSs. This script runs in the
-# CI to build the CJS and ESM libs.
-(
-  rm -rf ./lib && \
-  deno task build:intermediary && \
-  deno task build:esm && \
-  deno task build:cjs && \
-  rm -rf ./lib/intermediary
-)
