--- scripts/lib_builder/build_intermediary
@@ -1,11 +1,0 @@
-#!/bin/bash
-
-# This script only exists to run Deno in different OSs. This script runs in the
-# CI to build the CJS and ESM libs.
-(
-  deno fmt \
-  && deno run -A ./scripts/lib_builder/build_esm_lib.ts \
-    --debug \
-    --workspace=./lib/intermediary \
-    --copy-files=./src
-)
