--- scripts/lib_builder/deps.ts
@@ -1,6 +1,0 @@
-export {
-  emptyDirSync,
-  ensureDirSync,
-  walk,
-} from "https://deno.land/std@0.191.0/fs/mod.ts";
-export { copySync } from "https://deno.land/std@0.191.0/fs/copy.ts";
