--- scripts/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "allowImportingTsExtensions": true
-  },
-  "include": [
-    "./**/*.ts"
-  ]
-}
